Share session display name helper
